+++ rekorder/lib/player.py
@@ -1,7 +1,7 @@
 
 from .cassette import Cassette
 from .device import Device
-from .recorder import Recorder
+from .recorder.device import RecordingDevice
 from .what import What
 
 import importlib
@@ -12,13 +12,13 @@ class Player:
   '''
 
   def __init__(self, *args, **kwargs):
-    self._cassette = Cassette(*args, **kwargs)
+    self._recording_medium = Cassette(*args, **kwargs)
 
   def describe(self):
     '''Describe a recording.
     '''
 
-    for track in self._cassette.tracks:
+    for track in self._recording_medium.tracks:
       print(track.title)
 
       if not len(track.tunes):
@@ -28,7 +28,83 @@ class Player:
       for tune in track.tunes:
         print("  {}".format(tune.device))
 
-        if isinstance(tune.device, Recorder):
-          if self._cassette.recorder:
+        if isinstance(tune.device, RecordingDevice):
+          if self._recording_medium.recorder:
             raise Exception("Too many recorders")
-          self._cassette.recorder = tune.device
+          self._recording_medium.recorder = tune.device
+
+  def playback(self):
+    '''Play back a recording.
+    '''
+
+    # Fetch and playback the header track.
+    # This will intialize the recorder for playback validation.
+    rval = self._playback_header()
+
+    # Fetch and playback the entry track.
+    # This will invoke the entry point recorded by RecordingBegin.
+    rval = self._playback_body(rval)
+
+    # Playback of the entry track will have taken us through the
+    # recording & exit tracks and left us at the trailer.
+    rval = self._playback_trailer(rval=rval)
+
+  def _playback_header(self):
+    '''Playback the header to configure the workspace state and Recorder.
+    '''
+
+    track = self._recording_medium.track_manager.tracks.current_track
+    assert track.title == 'header'
+
+    rval = None
+
+    for tune in track.tunes:
+      print("  {}".format(tune.device))
+
+      if isinstance(tune.device, RecordingDevice):
+        self._engage_validation(tune.device.recorder)
+
+      rval = tune.playback(rval=rval)
+
+    track = self._recording_medium.track_manager.next
+
+    return rval
+
+  def _playback_body(self, rval):
+    '''Playback the entry, recording and exit tracks.
+    '''
+
+    track = self._recording_medium.track_manager.tracks.current_track
+    assert track.title == 'entry'
+
+    print("")
+    print(track.title)
+
+    if track.size() != 1:
+      raise Exception("There can be only one.")
+
+    tune = track.next_tune()
+    print("  {}".format(tune.device))
+    rval = tune.playback(rval=rval)
+
+    return rval
+
+  def _playback_trailer(self, rval):
+    '''Playback the trailer to restore workspace state.
+    '''
+
+    track = self._recording_medium.track_manager.tracks.current_track
+    assert track.title == 'trailer'
+
+    for tune in track.tunes:
+      print("  {}".format(tune.device))
+      rval = tune.playback(rval=rval)
+
+    return rval
+
+  def _engage_validation(self, recorder):
+    if self._recording_medium.recorder:
+      raise Exception("Too many recorders")
+    self._recording_medium.recorder = recorder
+    recorder.recording_medium = self._recording_medium
+    recorder.mode = What.VALIDATE