+++ rekorder/lib/tune.py
@@ -1,7 +1,10 @@
 
+import copy
+
 from . import util
 from .timestamp import Timestamp
 from .what import What
+from .when import When
 
 
 class Tune:
@@ -32,6 +35,8 @@ class Tune:
 
   @staticmethod
   def recordable_data(obj):
+    '''Used by Cassette when recording `obj` to the output file.
+    '''
 
     notes = {}
     for key, value in obj.notes.items():
@@ -59,14 +64,24 @@ class Tune:
       r.update({'when': obj.when})
     return r
 
-  def __init__(self, device, notes, when=None, timestamp=None, mode=What.RECORD):
+  def __init__(self, device, notes, when=When.NA, timestamp=None, mode=What.RECORD):
 
     self.mode = mode
     self.device = device
-    self.notes = notes
     self.when = when
     self.timestamp = timestamp if timestamp else Timestamp()
+
+    # Bugfix: If mutable parameters change during function invocation
+    #         MethodParameters recorded the new/changed value for both before
+    #         and after.
+    # We need to make a deep copy of the notes in case they change after this.
+    # For instance, if the notes represent function parameters it is possible
+    # that they could change during function invocation.
+    self.notes = copy.deepcopy(notes)
 
     from .device import Device
     if self.device and not isinstance(self.device, Device):
       raise Exception("[{}] is must be a Device".format(self.device))
+
+  def playback(self, *args, **kwargs):
+    return self.device.playback(*args, **kwargs)